requirements.txt
# absl-py==0.9.0
# atomicwrites==1.3.0
# attrs==19.3.0
# blackhc-laaos==0.1.2
# blackhc-notebook== 2.0.2
# blackhc-progress-bar==1.0.2
# boto==2.49.0
# boto3==1.12.12
# botocore==1.15.12
# click==7.1.1
# cycler==0.10.0
dataclasses==0.6
datasets==1.0.2
# dill==0.3.2
# docutils==0.15.2
# filelock==3.0.12
# gensim==3.8.1
# gitdb==4.0.2
# gitpython== 3.1.0
# google-auth==1.11.3
# google-auth-oauthlib==0.4.1
# grpcio==1.27.2
# h5py==2.10.0
# idna==2.9
# importlib-metadata==1.5.0
# jmespath==0.9.5
# joblib==0.14.1
# jsonpatch==1.25
# jsonpointer==2.0
# keras==2.3.1
# keras-applications==1.0.8
# keras-preprocessing==1.1.0
# kiwisolver==1.1.0
# llvmlite==0.34.0
# markdown==3.2.1
matplotlib==3.1.3
# more-itertools==8.2.0
# ninja==1.9.0
# nlp==0.4.0
nltk==3.4.5
numba==0.51.2
# numpy==1.18.1
# oauthlib==3.1.0
# olefile==0.46
overrides==2.8.0
pandas==0.25.3
# pillow==7.0.0
# pluggy==0.13.1
# protobuf==3.11.3
# psutil==5.7.0
# py ==1.8.1
# py-cpuinfo==5.0.0
# pyarrow==1.0.1
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycparser==2.19
# pyparsing==2.4.6
# pytest==3.10.0
# pytest-benchmark==3.2.3
# python-dateutil==2.8.1
# pytz==2019.3
# pyyaml ==3.12
# pyzmq==19.0.0
# regex==2020.2.20
requests==2.23.0
# requests-oauthlib==1.3.0
# rsa==4.0
# s3transfer==0.3.3
# sacremoses==0.0.38
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.10.0
sentencepiece==0.1.85
# seqeval==0.0.12
# setuptools==45.2.0
# six==1.14.0
# smart-open==1.9.0
# smmap==3.0.1
tabulate ==0.8.6
tensorboard==2.1.1
tensorboardx==2.0
torchfile==0.1.0
# torchvision==0.4.0
# tornado==6.0.3
tqdm==4.43.0
transformers==3.1.0
# typing==3.6.6
urllib3==1.25.8
# visdom==0.1.8.9
# websocket-client==0.57.0
# werkzeug==1.0.0
# wheel==0.34.2
# xxhash==2.0.0
# zipp==3.1.0
